% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_diffusion.R
\name{extract_diffusion}
\alias{extract_diffusion}
\title{Extract the Latitudinal Diffusion of a Crop}
\usage{
extract_diffusion(database, direction = c("north"))
}
\arguments{
\item{database}{A data.frame containing the crop database.}

\item{direction}{A string stating the direction (i.e., north).}
}
\value{
A tibble that has the frontier of a crop through time using 100 year time bins.
}
\description{
Extract the diffusion of a crop by northern latitude
}
