% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_earliest_crop.R
\name{extract_earliest_crop}
\alias{extract_earliest_crop}
\title{Extract Earliest Date for Crop by Province}
\usage{
extract_earliest_crop(database)
}
\arguments{
\item{database}{A data.frame containing the crop database.}
}
\value{
A tibble that has the earliest date for each province.
}
\description{
Extract earliest date for a crop by province
}
