% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_GDD.R
\name{extract_GDD}
\alias{extract_GDD}
\title{Extract GDD from WorldClim 2.0}
\usage{
extract_GDD(sites)
}
\arguments{
\item{sites}{A data.frame with an ID and sfc_Point geometry.}
}
\value{
A tibble with growing degree days for each ID.
}
\description{
Extract GDD from WorldClim 2.0 for 1970 to 2000 at Site Locations
}
