% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{global_gdd}
\alias{global_gdd}
\title{Accumulated Growing Degree Days.}
\format{
An object of class \code{RasterLayer}.
}
\source{
\url{https://sage.nelson.wisc.edu/data-and-models/atlas-of-the-biosphere/mapping-the-biosphere/ecosystems/growing-degree-days/}
}
\usage{
global_gdd
}
\description{
A \code{raster} containing global accumulated growing degree days for the year with a base temperature of 5°C.
}
\keyword{datasets}
