% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maizeDB}
\alias{maizeDB}
\title{North American Southwest Ancient Maize Database.}
\format{
An object of class \code{tibble}.
}
\source{
\url{https://doi.org/10.1038/s41597-022-01118-7}

\url{https://doi.org/10.48512/XCV8459173}

\url{http://en.ancientmaize.com/}

\url{https://www.canadianarchaeology.ca/}
}
\usage{
maizeDB
}
\description{
A \code{tibble} containing calibrated radiocarbon data for ancient maize macrosamples across the North American Southwest.
}
\references{
Bocinsky, R. Kyle, Darcy Bird, and Erick Robinson, (2020). Compendium of R code and data for p3k14c: A synthetic global database of archaeological radiocarbon dates. Accessed on September 12, 2022. https://github.com/people3k/p3k14c

Bird, D., Miranda, L., Vander Linden, M. et al. p3k14c, a synthetic global database of archaeological radiocarbon dates. Sci Data 9, 27 (2022). https://doi.org/10.1038/s41597-022-01118-7

Blake, M., B. Benz, D. Moreiras, L. Masur, N. Jakobsen and R. Wallace. 2017. Ancient Maize Map, Version 2.1: An Online Database and Mapping Program for Studying the Archaeology of Maize in the Americas. http://en.ancientmaize.com/. Laboratory of Archaeology, University of B.C., Vancouver. Accessed on February 24, 2021.

Martindale, Andrew, Richard Morlan, Matthew Betts, Michael Blake, Konrad Gajewski, Michelle Chaput, Andrew Mason, and Pierre Vermeersch (2016) Canadian Archaeological Radiocarbon Database (CARD 2.1), accessed September 10, 2022.
}
\keyword{datasets}
