% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NASW_elevation}
\alias{NASW_elevation}
\title{Digital Elevation Model for the Southwestern United States and Mexico States.}
\format{
An object of class \code{RasterLayer}.
}
\source{
\url{https://github.com/jhollist/elevatr}
}
\usage{
NASW_elevation
}
\description{
A \code{raster} containing elevation for the southwestern United States (i.e., Arizona, Colorado, New Mexico, and Utah) and Mexico.
}
\keyword{datasets}
