% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{swus_mexico_states}
\alias{swus_mexico_states}
\title{Southwestern United States and Mexico States Polygons}
\format{
An object of class \code{sfc_MULTIPOLYGON}.
}
\source{
\url{https://github.com/adeckmyn/maps}

\url{https://www.arcgis.com/home/item.html?id=ac9041c51b5c49c683fbfec61dc03ba8}
}
\usage{
swus_mexico_states
}
\description{
A \code{MULTIPOLYGON} containing the state boundaries for the southwestern United States (i.e., Arizona, Colorado, New Mexico, and Utah) and Mexico.
}
\keyword{datasets}
