% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{swus_states}
\alias{swus_states}
\title{Southwestern United States Polygons}
\format{
An object of class \code{sfc_MULTIPOLYGON}.
}
\source{
\url{https://github.com/adeckmyn/maps}
}
\usage{
swus_states
}
\description{
A \code{MULTIPOLYGON} containing the state boundaries for the southwestern United States (i.e., Arizona, Colorado, New Mexico, and Utah).
}
\keyword{datasets}
