% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WorldClim_annual_gdd}
\alias{WorldClim_annual_gdd}
\title{Annual Accumulated Growing Degree Days from WorldClim.}
\format{
An object of class \code{RasterLayer}.
}
\source{
\url{https://www.worldclim.org/data/worldclim21.html}
}
\usage{
WorldClim_annual_gdd
}
\description{
A \code{raster} containing annual accumulated growing degree days for the year with a base temperature of 10°C and maximum temperature of 30°C.
GDD was calculated from tmin and max 30 seconds (~1 \ifelse{html}{\out{km<sup>2</sup>}}{\eqn{km^2}}).
}
\keyword{datasets}
