% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_GDD.R
\name{extract_GDD}
\alias{extract_GDD}
\title{Extract growing degree days}
\usage{
extract_GDD(gdd_raster, sites)
}
\arguments{
\item{gdd_raster}{A raster containing growing degree days.}

\item{sites}{A data.frame with an ID and sfc_Point geometry containing site locations.}
}
\value{
A tibble with growing degree days for each ID.
}
\description{
\code{extract_GDD()} extracts growing degree days from a user provided GDD raster dataset at site locations.
}
