% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_GDD.R
\name{extract_GDD}
\alias{extract_GDD}
\title{Extract annual accumulated growing degree days}
\usage{
extract_GDD(sites)
}
\arguments{
\item{sites}{A data.frame with an ID and sfc_Point geometry containing site locations.}
}
\value{
A tibble with growing degree days for each ID.
}
\description{
\code{extract_GDD()} extracts annual accumulated growing degree days from WorldClim 2.0 for 1970 to 2000 at site locations.
}
